add fix register

+++ src/main/resources/db/migration/V1__create.sql
@@ -33,7 +33,7 @@ CREATE TABLE groups (
 CREATE TABLE user_role (
     id BIGSERIAL PRIMARY KEY,
     user_id BIGINT NOT NULL,
-    role_id BIGINT NOT NULL,
+    role_id INTEGER NOT NULL,
     FOREIGN KEY (user_id) REFERENCES users(id) ON DELETE CASCADE,
     FOREIGN KEY (role_id) REFERENCES role(id) ON DELETE CASCADE
 );
